requirements.txt
django>=5.0
psycopg2-binary>=2.9
supabase>=2.0
python-dotenv>=1.0
django-cors-headers>=4.0
PyJWT>=2.8
requests>=2.31

# AI/ML Libraries
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2

# NLP & Text Analysis
nltk>=3.8
textblob>=0.17.1
vaderSentiment>=3.3.2

# Image & Audio Processing
pytesseract>=0.3.10
Pillow>=10.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1

# API & Data Processing
newsapi-python>=0.2.7
beautifulsoup4>=4.12.0
lxml>=4.9.0
groq>=0.4.0

# Task Queue & Caching
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
